python/requirements.txt
# ApexQuant Python 依赖

# ==================== 核心依赖 ====================
numpy>=1.24.0
pandas>=2.0.0

# ==================== 数据获取 ====================
akshare>=1.12.0
baostock>=0.8.9  # 多数据源备份
# xtquant  # 需要从迅投官网单独下载安装

# ==================== AI/大模型 ====================
openai>=1.10.0  # DeepSeek API 兼容
anthropic>=0.18.0  # Claude API

# ==================== 机器学习 ====================
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0

# ==================== 深度学习（可选）====================
torch>=2.1.0
# tensorflow>=2.15.0  # 如果需要使用 TensorFlow

# ==================== 强化学习 ====================
ray[rllib]>=2.9.0
gymnasium>=0.29.0  # 新版 Gym

# ==================== 可视化 ====================
matplotlib>=3.8.0
plotly>=5.18.0
mplfinance>=0.12.10
seaborn>=0.13.0

# ==================== 数据处理 ====================
scipy>=1.11.0
statsmodels>=0.14.0
ta-lib>=0.4.28  # 技术指标库（需要先安装 TA-Lib C 库）

# ==================== 时间序列预测 ====================
prophet>=1.1.5  # Facebook Prophet

# ==================== 工具库 ====================
requests>=2.31.0
websocket-client>=1.7.0
python-dotenv>=1.0.0  # 环境变量管理
pyyaml>=6.0.1
loguru>=0.7.2  # 日志库
tqdm>=4.66.0  # 进度条
click>=8.1.7  # 命令行工具

# ==================== 测试 ====================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# ==================== 代码质量 ====================
black>=23.12.0  # 代码格式化
flake8>=7.0.0  # 代码检查
mypy>=1.8.0  # 类型检查

# ==================== Jupyter ====================
jupyter>=1.0.0
ipywidgets>=8.1.0

# ==================== 其他 ====================
python-dateutil>=2.8.2
pytz>=2023.3
chinesecalendar>=1.9.0  # 交易日历（A股节假日）

